Refactor unique seeding into per-format modules behind a trait
Introduces FormatSeeder trait and splits unique_existing.rs into a
directory module with one file per sink format (parquet, delta, iceberg),
so each new format is a self-contained addition with no changes to the
dispatch site.
