+++ crates/floe-core/src/run/entity/unique_existing/iceberg.rs
@@ -1,0 +1,205 @@
+use std::collections::HashMap;
+use std::path::Path;
+
+use arrow::record_batch::RecordBatch;
+
+use crate::errors::RunError;
+use crate::io::storage::{object_store, Target};
+use crate::io::write::iceberg::metadata::{
+    latest_gcs_metadata_location, latest_local_metadata_location, latest_s3_metadata_location,
+};
+use crate::io::write::iceberg::{
+    load_glue_table_state, GlueIcebergCatalogConfig, ICEBERG_CATALOG_NAME, ICEBERG_NAMESPACE,
+};
+use crate::{check, config, io, FloeResult};
+
+use super::{seed_from_batches, FormatSeeder};
+
+pub(super) struct IcebergSeeder<'a> {
+    pub(super) target: &'a Target,
+    pub(super) cloud: &'a mut io::storage::CloudClient,
+    pub(super) resolver: &'a config::StorageResolver,
+    pub(super) catalogs: &'a config::CatalogResolver,
+    pub(super) entity: &'a config::EntityConfig,
+}
+
+impl FormatSeeder for IcebergSeeder<'_> {
+    fn seed(
+        &mut self,
+        unique_tracker: &mut check::UniqueTracker,
+        scan_cols: &[String],
+        rename_back: &HashMap<String, String>,
+    ) -> FloeResult<()> {
+        // When a Glue catalog is configured, the writer uses glue_target.table_location as the
+        // real table root (not target). Seed from the same location so duplicate detection is
+        // consistent with what was actually written.
+        if let Some(glue_target) = self.catalogs.resolve_iceberg_target(
+            self.resolver,
+            self.entity,
+            &self.entity.sink.accepted,
+        )? {
+            if glue_target.catalog_type == "glue" {
+                return seed_from_glue_iceberg(
+                    unique_tracker,
+                    &glue_target,
+                    self.resolver,
+                    self.entity,
+                    scan_cols,
+                    rename_back,
+                );
+            }
+        }
+
+        let store = object_store::iceberg_store_config(self.target, self.resolver, self.entity)?;
+
+        let metadata_location: Option<String> = match self.target {
+            Target::Local { base_path, .. } => {
+                latest_local_metadata_location(Path::new(base_path))?
+            }
+            Target::S3 {
+                storage, base_key, ..
+            } => {
+                let client = self.cloud.client_for(self.resolver, storage, self.entity)?;
+                latest_s3_metadata_location(client, base_key)?
+            }
+            Target::Gcs {
+                storage, base_key, ..
+            } => {
+                let client = self.cloud.client_for(self.resolver, storage, self.entity)?;
+                latest_gcs_metadata_location(client, base_key)?
+            }
+            Target::Adls { .. } => return Ok(()),
+        };
+        let Some(metadata_location) = metadata_location else {
+            return Ok(());
+        };
+
+        let runtime = tokio::runtime::Builder::new_current_thread()
+            .enable_all()
+            .build()
+            .map_err(|err| {
+                Box::new(RunError(format!("iceberg seed runtime init failed: {err}")))
+            })?;
+
+        let batches = runtime
+            .block_on(collect_iceberg_batches(
+                metadata_location,
+                store.file_io_props,
+                store.warehouse_location,
+                &self.entity.name,
+                scan_cols,
+            ))
+            .map_err(|err| Box::new(RunError(format!("iceberg seed failed: {err}"))))?;
+
+        seed_from_batches(unique_tracker, batches, rename_back)
+    }
+}
+
+fn seed_from_glue_iceberg(
+    unique_tracker: &mut check::UniqueTracker,
+    glue_target: &config::ResolvedIcebergCatalogTarget,
+    resolver: &config::StorageResolver,
+    entity: &config::EntityConfig,
+    scan_cols: &[String],
+    rename_back: &HashMap<String, String>,
+) -> FloeResult<()> {
+    let catalog_target = Target::from_resolved(&glue_target.table_location)?;
+    let store = object_store::iceberg_store_config(&catalog_target, resolver, entity)?;
+
+    let glue_cfg = GlueIcebergCatalogConfig {
+        catalog_name: glue_target.catalog_name.clone(),
+        region: glue_target.region.clone(),
+        database: glue_target.database.clone(),
+        namespace: glue_target.namespace.clone(),
+        table: glue_target.table.clone(),
+    };
+
+    let runtime = tokio::runtime::Builder::new_current_thread()
+        .enable_all()
+        .build()
+        .map_err(|err| {
+            Box::new(RunError(format!(
+                "glue iceberg seed runtime init failed: {err}"
+            )))
+        })?;
+
+    let glue_state = runtime
+        .block_on(load_glue_table_state(&glue_cfg))
+        .map_err(|err| {
+            Box::new(RunError(format!(
+                "glue get_table for iceberg seed failed: {err}"
+            )))
+        })?;
+
+    let Some(metadata_location) = glue_state.metadata_location else {
+        return Ok(());
+    };
+
+    let batches = runtime
+        .block_on(collect_iceberg_batches(
+            metadata_location,
+            store.file_io_props,
+            store.warehouse_location,
+            &entity.name,
+            scan_cols,
+        ))
+        .map_err(|err| Box::new(RunError(format!("glue iceberg seed failed: {err}"))))?;
+
+    seed_from_batches(unique_tracker, batches, rename_back)
+}
+
+// Uses table.scan().to_arrow() so the Iceberg reader applies positional and equality deletes
+// before returning rows — only live rows are seeded.
+async fn collect_iceberg_batches(
+    metadata_location: String,
+    catalog_props: HashMap<String, String>,
+    warehouse_location: String,
+    entity_name: &str,
+    scan_columns: &[String],
+) -> Result<Vec<RecordBatch>, iceberg::Error> {
+    use futures::TryStreamExt;
+    use iceberg::memory::{MemoryCatalogBuilder, MEMORY_CATALOG_WAREHOUSE};
+    use iceberg::{Catalog, CatalogBuilder, NamespaceIdent, TableIdent};
+
+    let mut props = catalog_props;
+    props.insert(MEMORY_CATALOG_WAREHOUSE.to_string(), warehouse_location);
+
+    let catalog = MemoryCatalogBuilder::default()
+        .load(ICEBERG_CATALOG_NAME, props)
+        .await?;
+
+    let namespace = NamespaceIdent::new(ICEBERG_NAMESPACE.to_string());
+    if !catalog.namespace_exists(&namespace).await? {
+        catalog.create_namespace(&namespace, HashMap::new()).await?;
+    }
+
+    let table_name = iceberg_table_name(entity_name);
+    let table_ident = TableIdent::new(namespace, table_name);
+
+    let table = catalog
+        .register_table(&table_ident, metadata_location)
+        .await?;
+
+    let scan = table.scan().select(scan_columns.iter().cloned()).build()?;
+    let batch_stream = scan.to_arrow().await?;
+    batch_stream
+        .try_filter(|b| std::future::ready(b.num_rows() > 0))
+        .try_collect()
+        .await
+}
+
+fn iceberg_table_name(entity_name: &str) -> String {
+    let mut out = String::with_capacity(entity_name.len());
+    for ch in entity_name.chars() {
+        if ch.is_ascii_alphanumeric() || ch == '_' || ch == '-' {
+            out.push(ch);
+        } else {
+            out.push('_');
+        }
+    }
+    if out.is_empty() {
+        "table".to_string()
+    } else {
+        out
+    }
+}
